==> data/data_scripts/02_select_random_pixels.R <==
# select random pixels from all pixel locations

# load libraries
library(dplyr)
library(rbeni)

# read in pixel location data
pixel_locations <- readRDS("data/pixel_locations.rds")

# which Koeppen-Geiger classes to select, see table below
classes <- c(11:16, 21:28)

# Filter on KG classes, and group by IGBP vegetation type
# sample N pixels
sampled_pixels <- pixel_locations %>%
  dplyr::filter(kg_class %in% classes) %>%
  group_by(igbp) %>%
  slice_sample(n = 2500) #20000

# save to disk
saveRDS(sampled_pixels, "data/sampled_pixels.rds")

# Legend linking the numeric values in the maps to the Köppen-Geiger classes.
# The RGB colors used in Beck et al. [2018] are provided between parentheses.
# 
# 1:  Af   Tropical, rainforest                  [0 0 255]
# 2:  Am   Tropical, monsoon                     [0 120 255]
# 3:  Aw   Tropical, savannah                    [70 170 250]
# 4:  BWh  Arid, desert, hot                     [255 0 0]
# 5:  BWk  Arid, desert, cold                    [255 150 150]
# 6:  BSh  Arid, steppe, hot                     [245 165 0]
# 7:  BSk  Arid, steppe, cold                    [255 220 100]
# 8:  Csa  Temperate, dry summer, hot summer     [255 255 0]
# 9:  Csb  Temperate, dry summer, warm summer    [200 200 0]
# 10: Csc  Temperate, dry summer, cold summer    [150 150 0]
# 11: Cwa  Temperate, dry winter, hot summer     [150 255 150]
# 12: Cwb  Temperate, dry winter, warm summer    [100 200 100]
# 13: Cwc  Temperate, dry winter, cold summer    [50 150 50]
# 14: Cfa  Temperate, no dry season, hot summer  [200 255 80]
# 15: Cfb  Temperate, no dry season, warm summer [100 255 80]
# 16: Cfc  Temperate, no dry season, cold summer [50 200 0]
# 17: Dsa  Cold, dry summer, hot summer          [255 0 255]
# 18: Dsb  Cold, dry summer, warm summer         [200 0 200]
# 19: Dsc  Cold, dry summer, cold summer         [150 50 150]
# 20: Dsd  Cold, dry summer, very cold winter    [150 100 150]
# 21: Dwa  Cold, dry winter, hot summer          [170 175 255]
# 22: Dwb  Cold, dry winter, warm summer         [90 120 220]
# 23: Dwc  Cold, dry winter, cold summer         [75 80 180]
# 24: Dwd  Cold, dry winter, very cold winter    [50 0 135]
# 25: Dfa  Cold, no dry season, hot summer       [0 255 255]
# 26: Dfb  Cold, no dry season, warm summer      [55 200 255]
# 27: Dfc  Cold, no dry season, cold summer      [0 125 125]
# 28: Dfd  Cold, no dry season, very cold winter [0 70 95]
# 29: ET   Polar, tundra                         [178 178 178]
# 30: EF   Polar, frost                          [102 102 102]
# 
# Please cite Beck et al. [2018] when using the maps in any publication:
#   
#   Beck, H.E., N.E. Zimmermann, T.R. McVicar, N. Vergopolan, A. Berg, E.F. Wood:
#   Present and future Köppen-Geiger climate classification maps at 1-km resolution,
# Nature Scientific Data, 2018.
